// src/types/user.ts
export type UserType = {
  id: number;
  username: string;
  first_name: string;
  last_name: string;
  email: string;
};

export type SinginFormType = {
  email: string;
  password: string;
};

export type Tokens = {
  access: string;
  refresh: string;
};
export type StaredUserType = {
  id: number;
  username: string;
  first_name: string;
  last_name: string;
  email: string;
};